Weightlift tracking
Created a "Player body". For now it is just a cube, further improvement/changes required.
The player body has 2 colliders that act as triggers: shoulder level and knee level. When an object that is tagged as Liftable enters the shoulder level collider, an exercise is being tracked. After the object reaches the lower, knee level collider, the exercise is considered to be completed and a method is being called to handle exercise completion.